Always normalize to properly search for buck link trees
Summary:
Target building in fbcode was pretty broken - passing in a non-specific target such as `//tools/pyre/...` used to always error that "target is not normalized or built", because it was searching for the link tree "buck-out/gen/tools/pyre/...#*link-tree".

Not seeing a good workaround for this (we really want to be looking in `buck-out/gen/tools/pyre/scripts/etc.`), normalizing all targets and looking for their link trees specifically seems the best way to go.

Not sure why normalization is an option exposed to the user in the first place. Is this something they care about? Why not normalize?

In the case that only part of the supplied buck target is built, we used to just continue on with whatever link trees we found and the user might not know that pyre was not being run on only some of their desired targets. Warn users and prompt for rebuilding in this case.

Reviewed By: dkgi

Differential Revision: D6461608

fbshipit-source-id: f661712fa6dab4cb56294f155dd4e5d0c7dc2532