Filter errors outside of given buck target
Summary:
"External" makes sense to apply to everything outside of the original directory in distillery, but when targets are provided, I imagine it doesn't matter what the original directory is (since you can call a target from as high up as you want).

This makes "external" accurate for targets, and filteres out external targets in inference.

Reviewed By: dkgi

Differential Revision: D6523003

fbshipit-source-id: 41ceeb667f46648dd94c3e33e40b78b77ea809e2