Determine target not built if link trees not found
Summary: Previously had this in place to check if a target had been built independently from whether there was an associated link tree (since some targets do not produce link trees). Realized this is buggy with target names that are the prefix of other target names, and this distinction is unncessary (if the user passes in a target that even when built provides no link trees, it's fine that they kepe getting prompted that the target might not be built).

Reviewed By: sinancepel

Differential Revision: D6546343

fbshipit-source-id: fd1e1a7742706295eccb50212ed3e4ef51bf7604