Use buck show-output to increase built detection accuracy
Summary:
If we are running 'buck targets' already, we may as well pass in the '--show-output' flag which will return to us both the targets as well as the location of buck output for that target.

This way, we can actually know what to look for when checking whether a target is built or not. (Previously, we were just re-prompting every time for obscure targets that don't have link trees and therefore that we don't know how to search for, so the user has to say "yes" to build them every time even if they are already built, if they want to keep passing in the general /... buck target.)

While this will only happen if we need to normalize, that's kind of perfect because if someone's passing in a specific target, we won't normalize, but we also aren't really worried that this is an obscure target that we can't find. If someone's passing in a general target, we will always need to normalize it, and only then do we run into these obscure link-tree-less targets anyhow.

Reviewed By: dkgi

Differential Revision: D6553265

fbshipit-source-id: b51ae8fff591e7048ab142c52aad3fb81847e4d2