use lockfiles for server
Summary:
Start server wasn't reliable because lockfiles could be left behind after a crash, etc.
that caused a lock file to be left behind, and our checks only checked for existence of a file, not whether
the file was locked. Replace the previous check using Lock_files, which tether files to a process, and release
a lock when the file's gone.

Reviewed By: dkgi

Differential Revision: D6569365

fbshipit-source-id: ae90f6d5f5c8a310ec20e819147dee4b47446dca