CL arguments override configuration targets and trees
Summary:
Because we have .pyre_configuration set, we can no longer run pyre more granularly in (for example) distillery. Passing in a link tree is meaningless because all of distillery is set as a link tree in the root config.

If arguments are passed in, ignore the config targets/link trees.

Reviewed By: sinancepel

Differential Revision: D6580650

fbshipit-source-id: 252284266a3cf0d0d60168bb34e111393fc8ccbd