don't re-parse dependencies in incremental check
Summary:
A TypeCheckRequest marked with check_dependencies=false only originates from deferring the
computation of unchanged dependencies. We were doing extra work by re-parsing dependencies and repopulating
the environment with them, so don't do that for dependencies.

The result is that typechecking when distillery/accounts/helper.py is changed, the new typechecking time is
12.76s (down from 20). This time is dominated by a few very large test files, here's the biggest offender:

```
2017-12-20 09:06:58 CHECK Checked `discover/tests/main_tests.py` in 8.747606s
```

Reviewed By: dkgi

Differential Revision: D6611753

fbshipit-source-id: 5355d27e1a0b70f40a0ed9503b5ceeb181d42070