don't choke on `typing.List` without parameter
Summary:
not handling this has some unintended consequences. E.g. if we try to
resolve a method signature that takes a list without parameter as input we
might try to check if an argument is a subtype of that list. We won't find the
primitvie type in the type order and throw... causing us to miss a whole
function :(

Reviewed By: shannonzhu

Differential Revision: D6609554

fbshipit-source-id: 5ab4ffce0af06d8c4e2f75f3bd3956f731251c22