don't typecheck on didSave requests by default
Summary:
Nuclide sends us didSave requests even though our initialize response tells it not to,
so add an option to not turn didSave requests into TypeCheckRequests and enable this option in the server.
I chose to keep the didSave parsing code in because external users might not want to enable watchman, in which
case it's important that we handle these didSave requests.

Reviewed By: dkgi

Differential Revision: D6612691

fbshipit-source-id: 02c189734d3ac55ead3f22c867edfd3515b7153e