only use each shared memory prefix once
Summary:
Noticed while working on the shared type environment that the shared memory can only
discern the different tables by the prefixes, and the reuse causes removals to remove all values for
tables that happen to share the prefix. So, use each Value in one shared memory map, and rename accordingly.

Reviewed By: dkgi

Differential Revision: D6618027

fbshipit-source-id: a7cafcf2e37020af2d0950d3a140cb786d15afcf