widen precondition when inferring bottom in the return type constraints
Summary:
This is an issue that keeps occurring, e.g. we get a
`int_set.update(set(unknown))`. We infer that the return type of set
constructor is `Set[Bottom]`. When we apply this to infer the constraints of
`update` we get a type error because it'll restrict the inputs to the type
`Set[Bottom]` which cannot be fulfilled. Tl;dr: always tighten postconditions, losen
preconditions ;)

Reviewed By: mhlakhani

Differential Revision: D6634917

fbshipit-source-id: 9d0b01aa0f14c978e38f2c160e3c9459a29e8bc4