Refactor parent path logic in stub creation
Summary: There was a lot of ugly splitting and comparing that is hard to understand without printing out what the paths, parents, and names look like. This logic is mainly there to sanity check these relationships and filter out stubbing for nested classes. Am hoping this is much easier to understand.

Reviewed By: sinancepel

Differential Revision: D6569554

fbshipit-source-id: 0b655e47c304865518b1e90c724d7581ca9577ad