Optionalize return types in conditionals
Summary:
When types are being joined, if one of the types is "None" we should be treating this as Optional Bottom rather than as void for the join.

It would be nice if we can eventually migrate away from the concept of void entirely... everything is Optional Bottom but we can print these always as "None".

Reviewed By: dkgi

Differential Revision: D6650146

fbshipit-source-id: ede8469574eb3cdfb3165616f27f8af5fb6ac07b