Remove concept of void entirely
Summary:
As discussed with Lukasz and Carl just before the holidays, Python doesn't have a meaningful concept of "void". Consolidating "void" and "None" to mean the same thing here -- the Expression identifier "None" is a special-cased identifier that maps to Optional[Bottom] when represented as a type. All "None"s are thus treated the same, and will all be printed in errors as "None".

This is consistent with Python's runtime behavior in that non-returning functions really do return the same thing as a function explicitly returning None. Removes a lot of confusion/complexity to treat them the same in the type system.

Reviewed By: dkgi

Differential Revision: D6669362

fbshipit-source-id: 72cb89c33f6ce2782d7381592b5a401ee3712bfc