fix super issue
Summary:
The annotated fold wasn't special casing super() calls, and since there's no call named super(),
this was being viewed as a constructor for an unknown class in said fold. Pipe through define information
to annotated fold in order to be able to see which class we're in, and use the annotation of the superclass.
Fixes the super tests.

Reviewed By: dkgi

Differential Revision: D6682094

fbshipit-source-id: 8cf2834ac9902ed5ce9d63e2be0c78a69a99732f