don't add function bodies into shared memory
Summary:
This diff itself might not be the way to go; but I believe the idea is important: Typechecking time
is affected significantly by the amount of data we put into shared memory.

This diff strips function bodies before adding function definitions to the shared memory.
The typechecking uses the source directly, which has all the defines in there, and the typechecker is only
interested in the signatures from other files. If we choose not to do this, I still think there's value
in modifying the environment reader to contain a ClassSignature and FunctionSignature (names pending) types.

Highlights of the performance changes:
```
2017-12-20 14:50:49,606 PERFORMANCE Initial shared memory size: 125mb
2017-12-20 14:50:46,779 PERFORMANCE Full environment built in 25.506103s.
2017-12-20 14:50:55,525 PERFORMANCE Inferred protocol implementations in 5.918190s
2017-12-20 14:51:59,599 PERFORMANCE Time elapsed for check: 104.723619s
```

Reviewed By: dkgi

Differential Revision: D6615564

fbshipit-source-id: b111ae9e144d286ca3d2eb63f0b9bc14ef9e8861