Revert before unshelving changes
Summary:
Otherwise, shelved changes will be automatically rebased onto the state with all pyre-fixmes removed. Removal of fixmes looks like it should be a temporary state that is just used for grabbing pyre output results, so merging = need to re-extract changes out afterward.

Brought this up on D6671195 but noticed it again when using it yesterday.

Reviewed By: dkgi

Differential Revision: D6693490

fbshipit-source-id: e1049af45a585339db8690af8c022cdb7d8df398