also support non existing attributes in `Annotated.Attribute`
Summary:
Up until now we were only using the `Attribute` module for attributes we know are declared in a class. This makes it more difficult to deal with the case when they are not (e.g. for missing ones): we're using a separate data structure for them even though we're essentially encoding the same information.

(Note: this ignores all push blocking failures!)

Reviewed By: shannonzhu

Differential Revision: D6691348

fbshipit-source-id: 2479a871b9b29a9e93879b3ce818a6e788e1e724