Only check sources under type_check_root
Summary:
brainhurt

Fully resolve symlinks; only analyze files that fall under the directory level containing .pyre_config.

Still not sure how to run typeCheckService without printing a lot of output (for typeCheckService_test). Tried something along the lines of P58892271, then changed to this but realized maybe I need to spawn a server to do it? In which case, maybe it's better to move it to server_test.

Brain hurts from a full day of symlink and server test debugging :'(

Reviewed By: sinancepel

Differential Revision: D6710105

fbshipit-source-id: f13ad53413bb3c8d205289d16261aeb3b3025b39