don't include python 2 stubs
Summary:
The new type order was crashing because of a class in a 2and3/ stub subclassing UserDict.UserDict,
which only exists in python 2 stubs. However, this class was defined under a if sys.version_info < (3, ) conditional,
so remove those statements since we explicitly don't support python 2.

I'm not sure where to put this removal is - I placed it under qualify, but it might warrant its own preprocessing pass.
Looking for comments on that.

Reviewed By: dkgi

Differential Revision: D6699008

fbshipit-source-id: 9f1a58d88888756e867cb329a074789f163a25d9