Pass non-normalized targets into buck build
Summary:
Some target aliases map to a huge number of normalized targets (such as //one_world/...). This is crashing `buck build` when called on the entire list of targets.

Call build on the user-defined target when any subtarget is not built and the user prompts with "yes", instead of on all the normalized subtargets of that target.

Reviewed By: dkgi

Differential Revision: D6729142

fbshipit-source-id: ef2ba7216b84330a04373f46f39a59810d3d693d