RFC: split class and alias registration from type order building
Summary:
This diff sets up the groundwork for splitting the type order. It's a bit of a strugglebus,
as splitting the registration of types from the connecting of the type order causes the issue
of us erroring violently when we don't understand an alias.

Previously, we'd just create an unused class and move on, but the failure mode here is worse.Additionally, I'm not sure if this is a good split into phases.

Questions:
1. What is the ideal behavior when we encounter a superclass which isn't registered when register_type is running? I suspect this will happen for a while because our alias analysis isn't perfect.
2. Does the structure look OK on a high level? These are going to lead to some pretty destructive changes to the type order, would like to get the structure right in the first try.

Reviewed By: dkgi

Differential Revision: D6701555

fbshipit-source-id: d2c4b1ddbe01f633a03aa9f694900de31454e7b3