move registration of unknown superclasses to last phase
Summary:
This solves the problem of minting classes for aliases before we know better. As an example, consider

```
class _Collection:
      ...

Collection = _Collection

AbstractSet(Collection)
```

If we register AbstractSet's superclasses before doing the alias registration, we'll miss the AbstractSet -> _Collection edge,
because we'll create a new class not connected to the other Collection, and thus cause type errors when a _Collection is
expected.

Reviewed By: dkgi

Differential Revision: D6715917

fbshipit-source-id: c72c560dcd5dae6539cab74a58854a5863106a8a