Support `...` as the `Ellipsis` object in Python
Summary:
Python3 added `...` as an alias for the `Ellipsis` object:
```
>>> print(...)
Ellipsis
```
Hence, it's legal to use `...` as an identifier. For example:
```
a = ...
if a is ...:
   pass
```
This is a pattern used by the aiohttp-2.1.0 library, where `...` is
used by the Application class to mark an "uninitialized" field.

Add code in the parser to support such an usage of the
`Ellipsis`. Also remove ad-hoc code previously used to handle the
`ELLIPSES` token, to avoid a couple of reduce/reduce conflicts in the
grammar that were rendering two rules useless:

```
File "parser/pythonParser.mly", line 1444, characters 4-15:
Warning: production parser_pythonParser_separated_list(COMMA,define_parameter) -> expression option(annotation) EQUALS ELLIPSES is never reduced.
File "parser/pythonParser.mly", line 1395, characters 4-12:
Warning: production subscript -> ELLIPSES is never reduced.
```

Reviewed By: shannonzhu

Differential Revision: D6762370

fbshipit-source-id: a34867c9101fdc9dcde819ddc9d7e6f30522585e