explicitly handle static methods and attributes in fold
Summary: Previously we would ignore class names when traversing an access. E.g. in `some.Class.X` we would traverse the access and then find that `some.Class.X` is either an attribute or a function we know about. This makes it impossible for us to resolve generated class attributes (with `class_property`) so I've rearranged this to actually check if the current access is a class and if it is pass on its meta variable. In the example above this would see `some.Class`, realize it's a class that we know about, pass on `typing.Type[some.Class]` and then resolve methods and attributes on that.

Reviewed By: sinancepel

Differential Revision: D6754174

fbshipit-source-id: 282eeb366f103d8b7f84ca7d946c1a11f1767e9e