WIP: hackily merge multiple source directories in check
Summary:
Hack alert! This diff allows pyre to be reasonably performant in the case where pyre check
is called with a buck target spanning tens or hundreds of targets. The gist is that we create a .pyre/linked directory
in cwd (where cwd is the directory with a .pyre_configuration) and merge all source directories into there
in the case that we have multiple source directories.

This is WIP, so there are non-production-ready aspects. Putting this out there to get feedback on whether this
is a reasonable direction to take.

Reviewed By: dkgi

Differential Revision: D6754745

fbshipit-source-id: f4108046b2106a6596d4b6307e5fa817b6431698