double heap size and increase index table sizes
Summary:
We were crashing while analyzing the target
```
petra/tools/shard_tools:petra_shard_tools
```

because the total environment is larger than 1 GB, and has over 220k functions and 110k globals, resulting in
hitting the limit of 2^19 entries on the index table. Increase the upper limits to be able to build this target.

Reviewed By: dkgi

Differential Revision: D6776508

fbshipit-source-id: 4b30ffb2081e09f4c15b13dc3f4cb05e65629df0