refrain from using hd_exn in join and meet
Summary:
Hd_exn was causing crashes for the petra/tools/shard_tools:petra_shard_tools target due to a type
with zero successors. Although this is a legitimate issue we should fix in the typechecker, causing a crash
seems extreme. Default to reasonable values for join and meet, and log the issue.

Reviewed By: dkgi

Differential Revision: D6777092

fbshipit-source-id: 1f4eb7a48ec295e4ced1eac7eb9c7a7afb893a39