do not add non-existent supertypes to the type order
Summary:
This diff is quite useful for parallelization, but it makes the pyre errors on master
go from 83 to 72, indicating that there's a loss of precision when we don't know where a superclass
originates from.

The difference comes down (now that we have a choice) whether we want to keep around supertypes which
are never declared. I err on the side of not registering such supertypes. Reasoning is that if we don't
have a declaration of the supertype, we don't have any useful analysis, other than "You returned a C() when
an int was expected". The benefit is that this with this change, we can parallelize the registration of superclasses,
because we'll have all the type keys coming in. Open to other opinions!

Reviewed By: dkgi

Differential Revision: D6763776

fbshipit-source-id: 23d709ea0b0aabe0fda448da4ca7bf12cbdb779b