Scuba log per-file and aggregate type coverage
Summary:
We are currently logging type coverage statistics per source file checked. The idea was to aggregate these by "start time", a unique ID, to be able to track type coverage over a project we're interested in.

However, when check is run but not complete (error or user aborts), we get an incomplete set of samples that when summed can be arbitrarily lower than the true coverage numbers. I believe we should also be turning off this logging for incremental. Going to look into this further and create a dashboard for projects we're interested in.

Reviewed By: sinancepel

Differential Revision: D6795123

fbshipit-source-id: 1ac19e03abbf71db1a2c5b0e3c9d367acc6cc982