Tail server error logs in wrapper
Summary:
Previously, running incremental mode for the first time logged only "Initializing server" during the entire initial full check of the code, making it look like something was hanging or wrong.

{F119220802}

The call to incremental doesn't write the client's typecheck logging to stderr like a direct call to check, since it goes through the spawned server. Might have been nice to logic some of this logging in the client for parallel guarantees, but for now, tail the server logs to offer the same logging in incremental runs so the user knows the typechecker is hard at work.

{F119220771}

Reviewed By: dkgi

Differential Revision: D6783393

fbshipit-source-id: 6e701cac2fe96f31ca9c9e1f630a9c6735fa509b