Send project-root flag to client in all commands
Summary:
The type-check-root (soon to be renamed project-root) was only added to Check and Incremental commands; the server needs to know about this in the configuration that it is passed when started, otherwise incremental commands run on that server will be set with a project-root-less configuration.

This defaults to '/', meaning all files will be analyzed under incremental.

Reviewed By: dkgi

Differential Revision: D6813047

fbshipit-source-id: c23ce524efd383bd0840a5bf2f918c68276197e0