Option to use shared memory reader
Summary:
With some changes, we can finally use the shared memory reader, paving the way for parallelization.

Hide this between an environment variable for now - this will go away later, but there's inconsistencies in the
errors being reported, so they will be fixed before we only keep the shared memory reader around.

Reviewed By: dkgi

Differential Revision: D6825316

fbshipit-source-id: c3948bacafbd3810442c3c4a41c43108c6e717be