isolate pyre check shared source directories
Summary:
Noticed that if we called pyre with different targets in fbcode, a check would
eat up the server's files by removing everything from shared_source_directory. Suffix the
shared source directories for a standalone check with the PID in order to eliminate this class
of errors, clean up the directory after the check is done, and add a test.

Reviewed By: dkgi

Differential Revision: D6836885

fbshipit-source-id: 2dab4f2c179dda08c5db0532a887eb689578d5eb