add and ignore type-check-root in watchman
Summary:
The pyre watchman client was crashing instantly because it was being passed a type-check-root flag.
Take and ignore it for now. An alternative solution here is making sure that we don't pass the type-check-root flag
to watchman. I'm putting this diff up here because of the following question: Does it make sense to have a pyre
mode where we only watch updated files which come from within fbcode/, or should we always listen to updates
from third-party2 as well to pick up on dependency changes?

Reviewed By: dkgi

Differential Revision: D6838703

fbshipit-source-id: ed5376fd46ec689fb4f1de9604c4058f737f6623