set boundaries for bucket multiplier
Summary:
The overhead of dispatching jobs becomes very significant when the number of jobs per bucket falls below 20. This diff sets a lower bound on the job size for each process.

The end result is that a pyre check goes from 45s to 35s for the one_world/deployment target.

Reviewed By: dkgi

Differential Revision: D6910141

fbshipit-source-id: 3a63c7327923be71253cbda9b475e0b2e1edadb5