add a pyre-killall command.
Summary:
This is an experimental diff that creates a new pyre-killall script. I'm not convinced that this
is the way to go, but putting this here to start a discussion on how to best do "kill all the pyres".

What this does:
1. Read the configuration to get the absolute path to the pyre binary.
2. Iterate through /proc/ to see any processes which have the pyre binary path as its executable.
3. Send a sigkill to each such process.

Advantages:
- Straightforward, and relatively self contained implementation (still uses switch_root to pick up on the binary via .pyre_configurations).
- Does no harm, such as killing off random processes that happen to have "main.native" as its name.
- Doesn't carry the baggage of adding more logic to pyre.py, such as existing target-normalizing code that we want to avoid here. (Removing the code would be even more complexity in the wrapper script)>

Disadvantages:
- Can't kill pyre processes from previous versions.
- Repeats a bit of code from pyre.py and infer.py. I'm not too worried about this since that's about 3 loc, and the
whole script is less than 40.

Questions:
Is this the way to go? We could also merge this into the Kill command, with caveats laid out above.

What's the best way to test this? This script is very heavy in syscalls. Should I write a unit test that mocks these calls and lays out a fake "/proc", or is it better to create a sandcastle job that verifies no pyre process remains after pyre-killall has been run?

Reviewed By: dkgi

Differential Revision: D6935115

fbshipit-source-id: 81efe902b745307be8d3bf030c927b51b70aef0c