special case `typing.Optional` in attribute error dequalification
Summary:
we do dequalification of types before we generate the error messages.
Since we special case the optional type we cannot reliably recognize it as such
during the formatting phase if we dequalify it.

E.g. with

```
from typing import Optional
...
def foo(o: Optional[int]):
  o.attribute  # undefined
```

We'll first get the error `UndefinedAttribute { typing.Optional; ... }`,
dequalify it to make it match the imports to get `UndefinedAttribute { Optional; ... }`
and then try to special case the formatting for `typing.Optional` which doesn't
work anymore.

Reviewed By: sinancepel

Differential Revision: D6955793

fbshipit-source-id: e33996caf89a24a15455bb38f3fc86422dfc856c