refactor shared_source_directories out
Summary:
Self-explanatory. The interesting decision is where to specify the shared source directory name
for checks, which is different from the server one. Writing a function to generate one with an optional
suffix, and doing the selection during init seemed reasonable. Let me know if it's adding too much complexity.

Reviewed By: dkgi

Differential Revision: D6952455

fbshipit-source-id: 48cd8a16fb1c783e3255c8c7ead1591647f3b861