also register signal handlers in thread
Summary:
Signal handlers need to be registered for each thread - I had previously forgotten to do this, causing us to
not call stop_server when in the request_handler_thread, and leaving artifacts behind. This especially manifests itself when
the server is being initialized (before we get in the state where we wait for a request in select).

Differential Revision: D6953806

fbshipit-source-id: c23d0876e39f35e164f4a7cc3899ba8bdcc8c42b