renamed variables shadowed by imports
Summary:
looking at an access path `a.b.c()` right now we often don't know
whether `a` is a local variable (and the subsequent accesses attribute and
method accesses respectively), or if `a.b.c()` is a function that we're
importing from some module.

Things get doubly tricky if we miss certain definitions (e.g. cython right
now).

Longer term I'd like to be able to track modules in the backend (see T26023905)
and this should get easier. For now we should be able to get rid of most of
the special logic by renaming shadowed variables as to avoid confusion.

Reviewed By: shannonzhu

Differential Revision: D6978301

fbshipit-source-id: cf2ea048c146c2a83023da37f6a1f08958151e95