use hashmaps instead of excessive matching in Type.create
Summary:
Context, i.e. how I found this, because I know Dominik doesn't like unsystematic performance changes:
I noticed that pyre-performance-analysis got slower in D6929349 compared to the diff it's stacked on, indicating that that's a bottleneck.
Conjectured that this is because ocaml pattern matching is compiled down to if/elifs rather than jump tables for strings.
Implemented the change to use a hashtable as a way to restrict the number of matches being made.

Results:
Previously: 258.9s.
After: 190.2s.

Reviewed By: dkgi

Differential Revision: D6985126

fbshipit-source-id: 2cfaae6600803f0c2b5155fb443a23f786ad2663