don't add empty files when merging source directorieS
Summary:
We were overriding legitimate __init__ files with empty ones generated by buck when merging link trees.
This was causing nondeterminism in pyre check, which is not great. Check that the file we're adding is not empty
via stat(), and assume it is empty if we can't stat the file.

Reviewed By: dkgi

Differential Revision: D6993134

fbshipit-source-id: a6be0171db29b4e6a4623a1a4d47b4c2e6dcc696