add log rotation
Summary:
Initially went with this implementation instead of learning to use "logrotate" because this was simple enough.

The idea here is to support keeping the last N logs for your log_path so that we can handle crashes better.

Had a chat with mleogrande as I was finishing this diff up - it turns out that logrotate does have an option to keep the last N files
instead of being run as a cron job, and forcing the rotating of logs, which would accomplish this. Any opinions on the way to go?

Reviewed By: dkgi

Differential Revision: D7002071

fbshipit-source-id: 0da55f1f144a107f0b12f7ec454f942cfea0e1d4