Better error message for classvar declaration
Summary:
Forgetting to annotate class variables with typing.ClassVar is supercommon and potentially non-obvious, so add a note to error detail so it's clear when this is the problem.

Could be taken a step further by checking if there is an instance variable of the same name already declared in this class and providing the location.

Reviewed By: dkgi

Differential Revision: D6983953

fbshipit-source-id: d9bf739dfb46e1e1c4e5e9e26c0c69e8fe33cde8