Add module to represent ignores
Summary:
Keep track of more information about ignores than just ignored line and codes. This will allow us to 1. Warn on unused ignores and point to the location; 2. Differentiate between different kinds of ignores and treat them differently in metrics or in warnings.

Still to-do: Add this structure to the environment (we're just deconstructing it back down right now)

Reviewed By: sinancepel

Differential Revision: D7001229

fbshipit-source-id: 13d54777ef053c97fc09c8232ada590406843aa1