poll for deletion of server socket in stop command
Summary:
Mostly for user experience. The issue was that with $PYRE restart, the incremental command was being called
too quickly, which caused the incremental pyre to latch on to a closing server, and have a ugly looking "client exited, restart?"
message. This prevents that by polling for the socket deletion, so that the incremental check doesn't get in that
state.

Reviewed By: dkgi

Differential Revision: D7004747

fbshipit-source-id: 49c2b12f237117ef4563a45d59d0f384642e9f9a