lock server on the python side
Summary:
Ensures that only one server can be starting at the same time at a shared_source_directory/,
as well as ensuring that a server isn't running before merging directories in. This hierarchy of locks is not
going to deadlock because of a ordering of locks - a python wrapper lock can never be acquired after a server lock.

Differential Revision: D7016492

fbshipit-source-id: cdd16272a89579ce3d2e04bbeb1459afb8564ea6