use object instead of bottom for instantiating signatures
Summary:
When we try to instantiate a signature with Type.Bottom as the argument, we end up traversing the entire type order
without any new information - so, rename Type.Bottom -> Type.Object to avoid the unnecessary traversal.
Reduces typechecking time to 180s from 224s compared to the previous diff on my machine (there is a bit of instability in typechecking time, at least on my devserver,
so add a buffer of 10 seconds to verify statistical significance).

As a bonus, the single file type check of repute/service.py is significantly reduced - on my machine, the $PYRE_BINARY query "typecheck(repute/service.py)" is 3 seconds now, and I believe most of this is overhead due to deleting backedges/dependencies.

Reviewed By: shannonzhu

Differential Revision: D7036399

fbshipit-source-id: 56bfd90c39a9ac0f8dd3fa09d984239dbe015ce4