Error on unused pyre ignores
Summary:
Warn when pyre ignores or fixmes are not actually catching an error.

This was a requested feature from darkstorm team, and also makes sense because ignores can add clutter and also unintentionally hide type errors in the future if they are left behind. Can't really think of a reason someone might want a type ignore in their code even though there is no pyre error.

Reviewed By: dkgi

Differential Revision: D7045132

fbshipit-source-id: 8a850ce5cf4d84885e8b1b4f1fc721649d223425