Move ignore logic to Error module
Summary:
Moving this logic to Error;
Also moving internal logic around so that the flow is: 1 pass through errors that builds a lookup and filters ignored, 1 pass through ignores to lookup which are unused.

Thought about making defines in errors optional... I think it probably should be because there are enough cases where it isn't used. But seems relatively low priority

Reviewed By: dkgi

Differential Revision: D7076671

fbshipit-source-id: a38e80778f6d40a9538b9ba0af5604c2eeeead11