don't introduce `Unbound` when joining
Summary:
this seems to be the big false positive category here where we have an
assertion of the form `assert a is None or ...` and we do not have `a` in our
type environment (e.g. with fields we do not know about). Then joining the `or`
results in `Unbound` values.

Limiting unbounds to widening avoids this issue.

(Note: this ignores all push blocking failures!)

Reviewed By: sinancepel

Differential Revision: D7120369

fbshipit-source-id: 3bcde972591e87fab8e48054bde2324b78a494ab