add separate tags for development and release builds
Summary:
Some flags, such as unbox_closures and optimize(3) slow down the building
time but speed up the executable. Add support for these flags by copying _tags to _tags.development,
add _tags.release which contains the extra flags, and amend the Makefile to require the correct copying of
_tags to automate the process.

Question: I kept the _tags around as a default - does this make sense, or should we always force it to be generated, and add the _tags file to the .gitignore?

Reviewed By: dark

Differential Revision: D7112882

fbshipit-source-id: c29fcc221db5a24dcf59cff983ae2d6d5936384f