make or and and's right associative
Summary:
For context: This helps ease analyzing large boolean operator chains of the form
`a and b and c and d`, which gets parsed into `((a and b) and c) and d`. We analyze this as
```
check (assert a and b and c); (* Recursive asserts here *)
check f
```

Reordering the parsing to be right associative means that we'd now parse the above as
a and (b and (c and d)), which means we're only adding "assert a"; "assert b"; "assert c"; etc.

This specifically helps with `push/api/views.py`, which used to typecheck a single function with a large
or chain in 10 seconds. The time to typecheck that file becomes 1 second with this change.

Reviewed By: shannonzhu

Differential Revision: D7115319

fbshipit-source-id: c88c9afaa743080b28b6b6c8dedc99e787dd7abc