Generic types with no params should fill with Any
Summary:
Return types / parameters are sometimes annotated with parametric types with no parameters, which is expected to default to that type with "Any" in place of the missing parametric types.

Only adding this for commonly used builtin parametric types. Most of these have use-cases in fbcode/distillery where people are not passing in parameters. Not sure if we want to allow this kind of behavior for user-defined parametric types, but leaving it out for now.

Reviewed By: dkgi

Differential Revision: D7157815

fbshipit-source-id: cd954684e929a03b7393fa9bc62859bd79fae69c