Parse py2 type signatures with parameter annotations
Summary:
There are errors cropping up in hg_montor because we aren't picking up py2 annotations when the parameter annotations are not exactly '...'.

There are a variety of ways py2 annotations are formatted (with params set to '', '...', or to a list of types): https://fburl.com/biggrep/rknzf118

We still aren't picking up parameter annotations given in py2.

Reviewed By: dkgi

Differential Revision: D7172401

fbshipit-source-id: 101549edde282e009176d20bd412e96e9665e256