support contextmanagers
Summary:
Adds support for context manager, as well as setting up the infrastructure for handling more decorators in the future.

Copied what the stub under typeshed/2and3/contextlib.pyi does, which is take a callable taking an iterator, and use the iterator to construct a GeneratorContextManager.

Verified that this doesn't affect the typechecking of functions annotated as Iterator[int].

Reviewed By: dkgi

Differential Revision: D7175061

fbshipit-source-id: 0df541b2d5004515318f7b858f2524e23aa707f9