change type_check_request format to support both repopulated files and only typechecked files
Summary:
Context: I'm hoping to squash piled up server requests into one, where we take a union of the files
to repopulate and inferred files, and run it all as a single typecheck. The previous typecheck format didn't support this nicely.

Semantics shouldn't change in this diff.

Things to note:

- The server appends repopulate and only_check files at the start, and only_check = [] implies "check dependencies" right now. This will change in the future - here for consistency with the old code.

- You can signify an empty type check request as repopulate = [] and only_check = [].

Reviewed By: dkgi

Differential Revision: D7221073

fbshipit-source-id: 25ea8c193f656af0597ef1bd8912d8ccfb4a63f0