GC at high usage only
Summary:
We were garbage collecting with every request in e.g., instagram, even when the heap was quite underutilized.
Relax the requirement to only garbage collect when faced with a 50% or more utilization.

The number of slots used doesn't increase when we remove/add a key, but the heap utlization does. In other words, the garbage
collection doesn't change the number of slots at all. So only check the heap usage before GC'ing.

Reviewed By: dark

Differential Revision: D7244846

fbshipit-source-id: 5874ad9a4299ca2ba8eb576d64def7781a8bc9f9