don't add all backedges in a batch for now
Summary:
Partially revert the changes done to prepare for parallelizing the type order,
which include sequentially adding all backedges at the end. This is a stopgap measure. For the longer term,
when parallelizing the type order, I will also add support for incrementally adding backedges, which will mitigate this issue.

Shaves a second and a half off of server requests.

Reviewed By: dkgi

Differential Revision: D7264034

fbshipit-source-id: ae6fdaf15bacd0d7f5f17a8faa52e8769c657fb7