don't check integrity in incremental requests
Summary:
Checking the type order in incremental requests is quite expensive - over a second, which dominates the runtime
for single file type checks. Removing this incremental check for incremental requests and keeping it around for the initial
type order building removes this overhead, and brings down incremental check requests to more reasonable levels.

Reviewed By: dkgi

Differential Revision: D7265690

fbshipit-source-id: 66a328cd0784f40e8d0216e8f57231b7369f742b