fix qualification for annotations within type comments
Summary:
When parsing a signature annotation, we would treat the parsed string as Access via
Access.create in the parser. This doesn't work, because Access.create doesn't handle subscripts and
recursive accesses (to do so would require a parser). This is a quick-and-dirty solution that adds the annotation
as a string instead, forcing the typechecker to re-parse it at a later point and infer the correct structure.

Adds a test to verify that we're doing the right thing (tm).

Reviewed By: shannonzhu

Differential Revision: D7298226

fbshipit-source-id: f09dfd13806ef018b0f02bc8175d08347187aafd